requirements.txt
Django>=4.2.9,<4.2.10
djangorestframework>=3.14.0,<3.15
django-cors-headers>=4.0.0,<4.1
django-extensions>=3.2.1,<3.3
django-jazzmin>=2.6.0,<2.7
psycopg[c]>=3.1.8,<3.2
djangorestframework-simplejwt>=5.0.0,<5.3.1
django-silk>=5.0.3,<5.1
cryptography>=41.0.0,<=41.0.5
uvicorn[standard]>=0.27.0,<=0.27.0.post1
channels[daphne]==4.0.0
channels-redis==4.2.0
pycountry==23.12.11
django-map-widgets==0.4.2
geoip2==4.8.0
Faker==23.1.0
celery>=5.3.5,<5.3.6
python-magic==0.4.27
django-versatileimagefield==3.0
phonenumbers==8.13.30
django-phonenumber-field==7.0.2
pillow==9.4.0
drf-yasg==1.21.7
django-postgres-extra==2.0.8